# [ArrowStringArray] PERF: str.extract object fallback