```
       before           after         ratio
     [40482273]       [1788ffe6]
     <master>         <extract-single-group>
-         134±2ms          106±1ms     0.79  strings.Extract.time_extract_single_group('string', False)
-         359±5ms        110±0.8ms     0.31  strings.Extract.time_extract_single_group('arrow_string', False)

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```