@simonjayhawkins 2021-05-16
Have dumbed down the changes to just the materialization of the array. Iteration on a pyarow array is the main bottleneck here.

```
       before           after         ratio
     [40482273]       [79ff0d3e]
     <master>         <extract-single-group>
-         140±2ms          114±2ms     0.82  strings.Methods.time_extract('string')
-       134±0.9ms          107±2ms     0.80  strings.Extract.time_extract_single_group('string', False)
-         139±2ms        108±0.9ms     0.77  strings.Extract.time_extract_single_group('string', True)
-         356±3ms          118±2ms     0.33  strings.Extract.time_extract_single_group('arrow_string', False)
-        367±10ms          120±1ms     0.33  strings.Methods.time_extract('arrow_string')
-         360±2ms          112±1ms     0.31  strings.Extract.time_extract_single_group('arrow_string', True)

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```